# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Pathology Network API",
        "version": "1.0.0",
        "description": "Pathology Network Platform API v1. Errors use { error: { code, message, status, details } }; lists use { data, pagination: { next_cursor, limit } }."
    },
    "servers": [
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "paths": {
...
        "/branches/{branch}": {
            "get": {
                "operationId": "branch.show",
                "tags": [
                    "Branch"
                ],
                "parameters": [
                    {
                        "name": "branch",
                        "in": "path",
                        "required": true,
                        "description": "The branch ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`BranchResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BranchResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "branch.update",
                "tags": [
                    "Branch"
                ],
                "parameters": [
                    {
                        "name": "branch",
                        "in": "path",
                        "required": true,
                        "description": "The branch ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BranchRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`BranchResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BranchResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "branch.destroy",
                "tags": [
                    "Branch"
                ],
                "parameters": [
                    {
                        "name": "branch",
                        "in": "path",
                        "required": true,
                        "description": "The branch ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/departments/{department}": {
            "get": {
                "operationId": "department.show",
                "tags": [
                    "Department"
                ],
                "parameters": [
                    {
                        "name": "department",
                        "in": "path",
                        "required": true,
                        "description": "The department ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`DepartmentResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/DepartmentResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "department.update",
                "tags": [
                    "Department"
                ],
                "parameters": [
                    {
                        "name": "department",
                        "in": "path",
                        "required": true,
                        "description": "The department ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DepartmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`DepartmentResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/DepartmentResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "department.destroy",
                "tags": [
                    "Department"
                ],
                "parameters": [
                    {
                        "name": "department",
                        "in": "path",
                        "required": true,
                        "description": "The department ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/tests/{test}": {
            "get": {
                "operationId": "labTest.show",
                "tags": [
                    "LabTest"
                ],
                "parameters": [
                    {
                        "name": "test",
                        "in": "path",
                        "required": true,
                        "description": "The test ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`LabTestResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/LabTestResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "parameters"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "labTest.update",
                "tags": [
                    "LabTest"
                ],
                "parameters": [
                    {
                        "name": "test",
                        "in": "path",
                        "required": true,
                        "description": "The test ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LabTestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`LabTestResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/LabTestResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "labTest.destroy",
                "tags": [
                    "LabTest"
                ],
                "parameters": [
                    {
                        "name": "test",
                        "in": "path",
                        "required": true,
                        "description": "The test ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/packages/{package}": {
            "get": {
                "operationId": "package.show",
                "tags": [
                    "Package"
                ],
                "parameters": [
                    {
                        "name": "package",
                        "in": "path",
                        "required": true,
                        "description": "The package ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`PackageResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/PackageResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "tests"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "package.update",
                "tags": [
                    "Package"
                ],
                "parameters": [
                    {
                        "name": "package",
                        "in": "path",
                        "required": true,
                        "description": "The package ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PackageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`PackageResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/PackageResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "tests"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "package.destroy",
                "tags": [
                    "Package"
                ],
                "parameters": [
                    {
                        "name": "package",
                        "in": "path",
                        "required": true,
                        "description": "The package ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/price-lists/{priceList}": {
            "get": {
                "operationId": "priceList.show",
                "tags": [
                    "PriceList"
                ],
                "parameters": [
                    {
                        "name": "priceList",
                        "in": "path",
                        "required": true,
                        "description": "The price list ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`PriceListResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PriceListResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "priceList.update",
                "tags": [
                    "PriceList"
                ],
                "parameters": [
                    {
                        "name": "priceList",
                        "in": "path",
                        "required": true,
                        "description": "The price list ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PriceListRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`PriceListResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PriceListResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "priceList.destroy",
                "tags": [
                    "PriceList"
                ],
                "parameters": [
                    {
                        "name": "priceList",
                        "in": "path",
                        "required": true,
                        "description": "The price list ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BranchResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "branch_code": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "branch_type": {
                        "$ref": "#/components/schemas/BranchType"
                    },
                    "mrp_price_list_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "owner_type": {
                        "$ref": "#/components/schemas/BranchOwnerType"
                    },
                    "franchise_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "region_id": {
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/BranchStatus"
                    },
                    "nabl_certificate_no": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "nabl_valid_till": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "hfr_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "clinical_establishment_reg_no": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "address": {
                        "type": "string"
                    },
                    "pincode": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "longitude": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "phone": {
                        "type": "string"
                    },
                    "working_hours": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "additionalProperties": {}
                    },
                    "opened_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "branch_code",
                    "name",
                    "branch_type",
                    "mrp_price_list_id",
                    "owner_type",
                    "franchise_id",
                    "region_id",
                    "status",
                    "nabl_certificate_no",
                    "nabl_valid_till",
                    "hfr_id",
                    "clinical_establishment_reg_no",
                    "address",
                    "pincode",
                    "latitude",
                    "longitude",
                    "phone",
                    "working_hours",
                    "opened_at",
                    "created_at",
                    "updated_at"
                ],
                "title": "BranchResource"
            },
...
            "DepartmentRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 80
                    },
                    "signing_discipline": {
                        "$ref": "#/components/schemas/SigningDiscipline"
                    },
                    "report_order": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 1000
                    }
                },
                "required": [
                    "name",
                    "signing_discipline"
                ],
                "title": "DepartmentRequest"
            },
...
        },
        "responses": {
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}